Add unit test for GetIndexDefinitions()
There was no need for this being a functional test (the same is actually true for a bunch of the tests in MongoManager_EnsureIndexes_Tests).
