+++ rapidcore.mongo.sln
@@ -9,6 +9,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{91683FE2-6
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "functionaltests", "test\functional\functionaltests.csproj", "{76979851-7B58-4EB8-BE1B-B2FAB4D131CB}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "unittests", "test\unit\unittests.csproj", "{6249341E-B470-462E-89E9-CE76EFAB88AD}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -46,8 +48,21 @@ Global
 		{76979851-7B58-4EB8-BE1B-B2FAB4D131CB}.Release|x64.Build.0 = Release|x64
 		{76979851-7B58-4EB8-BE1B-B2FAB4D131CB}.Release|x86.ActiveCfg = Release|x86
 		{76979851-7B58-4EB8-BE1B-B2FAB4D131CB}.Release|x86.Build.0 = Release|x86
+		{6249341E-B470-462E-89E9-CE76EFAB88AD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6249341E-B470-462E-89E9-CE76EFAB88AD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6249341E-B470-462E-89E9-CE76EFAB88AD}.Debug|x64.ActiveCfg = Debug|x64
+		{6249341E-B470-462E-89E9-CE76EFAB88AD}.Debug|x64.Build.0 = Debug|x64
+		{6249341E-B470-462E-89E9-CE76EFAB88AD}.Debug|x86.ActiveCfg = Debug|x86
+		{6249341E-B470-462E-89E9-CE76EFAB88AD}.Debug|x86.Build.0 = Debug|x86
+		{6249341E-B470-462E-89E9-CE76EFAB88AD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6249341E-B470-462E-89E9-CE76EFAB88AD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6249341E-B470-462E-89E9-CE76EFAB88AD}.Release|x64.ActiveCfg = Release|x64
+		{6249341E-B470-462E-89E9-CE76EFAB88AD}.Release|x64.Build.0 = Release|x64
+		{6249341E-B470-462E-89E9-CE76EFAB88AD}.Release|x86.ActiveCfg = Release|x86
+		{6249341E-B470-462E-89E9-CE76EFAB88AD}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{76979851-7B58-4EB8-BE1B-B2FAB4D131CB} = {91683FE2-6739-437A-8BD9-76BDD333FD33}
+		{6249341E-B470-462E-89E9-CE76EFAB88AD} = {91683FE2-6739-437A-8BD9-76BDD333FD33}
 	EndGlobalSection
 EndGlobal
