+++ test/unit/unittests.csproj
@@ -1,0 +1,21 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp1.1</TargetFramework>
+    <RootNamespace>RapidCore.Mongo.UnitTests</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.0.*" />
+    <PackageReference Include="MongoDB.Bson" Version="2.4.*" />
+    <PackageReference Include="MongoDB.Driver" Version="2.4.*" />
+    <PackageReference Include="MongoDB.Driver.Core" Version="2.4.*" />
+    <PackageReference Include="xunit" Version="2.2.*" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.2.*" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\rapidcore.mongo.csproj" />
+  </ItemGroup>
+
+</Project>
